pyproject.toml
[tool.poetry]
name = "nonebot-plugin-updater"
version = "0.1.0"
description = "一款全新的检测已安装插件更新情况的插件"
authors = ["hanasa2023 <hanasakayui2022@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9, < 4.0"
nonebot2 = ">=2.3.3"
httpx = ">=0.27.2"
nonebot-plugin-alconna = ">=0.52.3"
toml = "^0.10.2"


[[tool.poetry.source]]
name = "mirrors"
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple/"
priority = "primary"


[tool.poetry.group.dev.dependencies]
mypy = "^1.11.2"
nonebot-adapter-onebot = "^2.4.5"
fastapi = "^0.115.0"
nonebot2 = { extras = ["fastapi"], version = "^2.3.3" }
nonebot-plugin-status = "0.8.0"
types-toml = "^0.10.8.20240310"

[tool.ruff]
ignore = ["E402", "F401"]

[tool.ruff.format]
quote-style = "single"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" },
]
plugins = []
plugin_dirs = [""]
builtin_plugins = []
